Moved configs to own files
