+++ addons/factions/gar/configs/GroundHolders.hpp
@@ -1,0 +1,20 @@
+class DOUBLES(Ground,CLASS(Holder_Base));
+GROUND_HOLDER(Helmet_Phase1_CG,"INF P1 Helm (CG)");
+
+GROUND_HOLDER(Helmet_Phase1_HowzerBlue,"INF P1 Helm ('Howzer') [Blue]");
+GROUND_HOLDER(Helmet_Phase1_HowzerBrown,"INF P1 Helm ('Howzer') [Brown]");
+
+GROUND_HOLDER(Helmet_Phase1_91st_Ponds,"INF P1 Helm ('Ponds')");
+
+GROUND_HOLDER(Helmet_Phase2_CG,"INF P2 Helm (CG)");
+GROUND_HOLDER(Helmet_Phase2_CG_Stone,"INF P2 Helm ('Stone')");
+
+GROUND_HOLDER(Helmet_Phase2_Fil,"INF P2 Helm ('Fil')");
+GROUND_HOLDER(Helmet_Phase2_FilSquad,"INF P2 Helm ('Fil's Squad')");
+
+GROUND_HOLDER(Helmet_Phase2_HowzerBlue,"INF P2 Helm ('Howzer') [Blue]");
+GROUND_HOLDER(Helmet_Phase2_HowzerBrown,"INF P2 Helm ('Howzer') [Brown]");
+
+GROUND_HOLDER(Helmet_Phase2_187th,"INF P2 Helm (187th Trooper)");
+
+GROUND_HOLDER(Helmet_Phase2_91st_Ponds,"INF P2 Helm ('Ponds')");
