+++ addons/factions/gar/configs/Helmets_P2.hpp
@@ -1,0 +1,94 @@
+class CLASS(Helmet_Phase2_Base);
+class CLASS(Helmet_Phase2_CG): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[CG] INF P2 Helm";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\CG_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
+
+class CLASS(Helmet_Phase2_CG_Stone): CLASS(Helmet_Phase2_CG)
+{
+    displayName = "[CG] INF P2 Helm ('Stone')";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\Stone_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+};
+
+class CLASS(Helmet_Phase2_Fil): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[KC] INF P2 Helm ('Fil')";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\Fil_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
+
+class CLASS(Helmet_Phase2_FilSquad): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[KC] INF P2 Helm ('Fil's Squad')";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\FilSquad_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
+
+class CLASS(Helmet_Phase2_HowzerBlue): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[KC] INF P2 Helm ('Howzer') [Blue]";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\HowzerBlue_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
+
+class CLASS(Helmet_Phase2_HowzerBrown): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[KC] INF P2 Helm ('Howzer') [Brown]";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\HowzerBrown_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
+
+class CLASS(Helmet_Phase2_187th): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[187th] INF P2 Helm";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\187th_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
+
+class CLASS(Helmet_Phase2_91st_Ponds): CLASS(Helmet_Phase2_Base)
+{
+    displayName = "[91st] INF P2 Helm ('Ponds')";
+    hiddenSelectionsTextures[] =
+    {
+        QPATHTOF(SUBCOMPONENT\data\helmets\phase2\Ponds_camo1_co.paa),
+        "\ls_armor_bluefor\helmet\gar\phase2\data\visor_co.paa"
+    };
+
+    EGVAR(custom_armor,isCustom) = TRUE;
+};
